push local changes

+++ include/stddef.h
@@ -3,9 +3,9 @@
 
 #define NULL ((void *)0)
 
-typedef unsigned long size_t;
-typedef long ptrdiff_t;
-typedef unsigned int wchar_t;
+typedef unsigned int size_t;
+typedef int ptrdiff_t;
+typedef unsigned short wchar_t;
 typedef long max_align_t;
 
 #define offsetof(type, member) ((size_t)&(((type *)0)->member))
